backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
python-dotenv==1.0.0
python-multipart==0.0.6
groq==0.4.2
httpx==0.27.2
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.2.0
python-docx==1.1.0
pymongo==4.6.1
bcrypt==4.1.2
